+++ services/sso/src/controllers/auth/loginController.ts
@@ -1,7 +1,7 @@
 import { Context, Middleware } from 'koa'
-import { client } from '../../models/db'
+import { client } from '../../db/client'
 import { UserLogin } from '../../schemas/auth/loginSchema'
-import { generateToken } from '../../utils/auth'
+import { generateToken } from '../../utils/jwtAuth'
 import { ForbiddenError, InvalidCredentials } from '../../utils/errors'
 import { checkPassword } from '../../utils/passwordHash'
 import { User, UserStatus } from '../../schemas/users/userSchema'
@@ -27,8 +27,8 @@ export const loginController: Middleware = async (ctx: Context) => {
   if (user.status !== UserStatus.ACTIVE) {
     throw new ForbiddenError('Only active users can login')
   }
-  const authToken = generateToken(user.id, '15m')
-  const refreshToken = generateToken(user.id, '7d')
+  const authToken = generateToken(user.id, 'auth')
+  const refreshToken = generateToken(user.id, 'refresh')
   ctx.status = 200
   ctx.body = { id: user.id, authToken, refreshToken }
 }