+++ services/sso/src/controllers/auth/registerController.ts
@@ -1,6 +1,6 @@
 import { Context, Middleware } from 'koa'
 import { UserRegister } from '../../schemas/auth/registerSchema'
-import { client } from '../../models/db'
+import { client } from '../../db/client'
 import { hashPassword } from '../../utils/passwordHash'
 import { generateId } from '../../utils/cuidGenerator'