+++ services/sso/src/controllers/tokens/refresh.ts
@@ -1,0 +1,10 @@
+import { Context, Middleware } from 'koa'
+import { generateToken } from '../../utils/jwtAuth'
+import { User } from '../../schemas'
+
+export const refresh: Middleware = async (ctx: Context) => {
+  const { id } = ctx.state.user as Pick<User, 'id' | 'role'>
+  const authToken = generateToken(id, 'auth')
+  ctx.status = 200
+  ctx.body = { authToken }
+}