+++ services/sso/src/middleware/authenticate.ts
@@ -1,34 +1,41 @@
 import { Context, Next } from 'koa'
-import jwt, { JwtPayload } from 'jsonwebtoken'
-import { ForbiddenError, InvalidCredentials } from '../utils/errors'
+import jwt from 'jsonwebtoken'
+import { InvalidCredentials } from '../utils/errors'
 import { appConfig } from '../config'
-import { client } from '../models/db'
-import { UserStatus } from '../schemas/users/userSchema'
+import { generateToken } from '../utils/jwtAuth'
+import { validateUserAndToken } from '../utils/authHelper'
 
 export const authenticate = async (ctx: Context, next: Next) => {
-  const authToken = ctx.request.body?.authToken
-  const authCookie = ctx.cookies.get('authToken')
-  const token = authToken || authCookie
+  const authHeader = ctx.request.headers.authorization
+  const authTokenFromCookie = ctx.cookies.get('authToken')
+  const authTokenFromBody: string | undefined = ctx.request.body?.authToken
+  let authToken
+  if (authHeader?.startsWith('Bearer ')) {
+    authToken = authHeader.slice(7)
+  } else {
+    authToken = authTokenFromBody ?? authTokenFromCookie
+  }
+  const refreshToken = ctx.cookies.get('refreshToken')
 
-  if (!token) {
-    throw new InvalidCredentials()
-  }
-  const { id } = jwt.verify(token, appConfig.jwtKey) as JwtPayload
+  try {
+    const user = await validateUserAndToken('auth', authToken)
+    ctx.state.user = user
+    await next()
+  } catch (error) {
+    if (error instanceof jwt.TokenExpiredError) {
+      if (authTokenFromBody || authHeader)
+        throw new InvalidCredentials('Token has expired')
+      const user = await validateUserAndToken('refresh', refreshToken)
+      const newToken = generateToken(user.id, 'auth')
+      ctx.cookies.set('authToken', newToken, {
+        httpOnly: true,
+        maxAge: appConfig.authCookieExpiration,
+      })
 
-  const userResult = await client.query(
-    'SELECT id, role, status FROM "user" WHERE id = $1',
-    [id]
-  )
-  const user = userResult.rows[0]
-  if (!user) {
-    throw new InvalidCredentials()
+      ctx.state.user = user
+      await next()
+    } else {
+      throw error
+    }
   }
-  if (user.status === UserStatus.BLOCKED) {
-    throw new ForbiddenError('The user is Blocked')
-  }
-  if (user.status !== UserStatus.ACTIVE) {
-    throw new ForbiddenError('Only active users can proceed')
-  }
-  ctx.state.user = user
-  await next()
 }