+++ web/usuaris/src/__mocks__/handlers.ts
@@ -49,6 +49,7 @@ export const handlers = [
           role: UserRole.ADMIN,
           createdAt: '2023/11/05 00:00:00.000',
           itineraryName: 'Backend Node',
+          deletedAt: null,
         },
         {
           id: '2',
@@ -58,6 +59,7 @@ export const handlers = [
           role: UserRole.REGISTERED,
           createdAt: '2023/11/06 00:00:00.000',
           itineraryName: 'Frontend React',
+          deletedAt: null,
         },
         {
           id: '3',
@@ -67,6 +69,7 @@ export const handlers = [
           role: UserRole.REGISTERED,
           createdAt: '2023/11/07 00:00:00.000',
           itineraryName: 'Fullstack Php',
+          deletedAt: null,
         },
         {
           id: '4',
@@ -76,6 +79,7 @@ export const handlers = [
           role: UserRole.ADMIN,
           createdAt: '2023/11/08 00:00:00.000',
           itineraryName: 'Frontend React',
+          deletedAt: null,
         },
         {
           id: '5',
@@ -85,6 +89,17 @@ export const handlers = [
           role: UserRole.ADMIN,
           createdAt: '2023/11/09 00:00:00.000',
           itineraryName: 'Frontend Angular',
+          deletedAt: null,
+        },
+        {
+          id: '6',
+          name: 'Paula Font',
+          dni: '12131415F',
+          status: 'BLOCKED',
+          role: UserRole.REGISTERED,
+          createdAt: '2023/10/10 00:00:00.000',
+          itineraryName: 'Data Science',
+          deletedAt: '2023/10/11 00:00:00.000',
         },
       ],
       { status: 200 }