+++ web/usuaris/src/constants/languages.ts
@@ -1,0 +1,19 @@
+export type Language = 'es' | 'cat' | 'en'
+
+export const languages = [
+  {
+    label: 'CAT',
+    value: 'cat',
+    id: 'cat',
+  },
+  {
+    label: 'ES',
+    value: 'es',
+    id: 'es',
+  },
+  {
+    label: 'EN',
+    value: 'en',
+    id: 'en',
+  },
+]