+++ web/usuaris/src/types/types.tsx
@@ -2,6 +2,7 @@ export enum UserStatus {
   ACTIVE = 'ACTIVE',
   PENDING = 'PENDING',
   BLOCKED = 'BLOCKED',
+  DELETED = 'DELETED',
 }
 
 export type TFilters = {
@@ -46,6 +47,7 @@ export type TUserData = {
   createdAt: string
   itineraryName: string
   role: string
+  deletedAt: string
 }
 
 export enum UserRole {